timer: Plug API into parent class.
