+++ lib/vpt/timer/timer-const.ts
@@ -17,24 +17,13 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-import { EventProxy } from '../../game/event-proxy';
-import { Player } from '../../game/player';
-import { ItemApi } from '../item-api';
-import { Table } from '../table/table';
-import { TimerData } from './timer-data';
-
-export class TimerApi extends ItemApi<TimerData> {
-
-	constructor(data: TimerData, events: EventProxy, player: Player, table: Table) {
-		super(data, events, player, table);
-	}
-
-	get X() { return this.data.vCenter.x; }
-	set X(v) { this.data.vCenter.x = v; }
-	get Y() { return this.data.vCenter.y; }
-	set Y(v) { this.data.vCenter.y = v; }
-	get Interval() { return this.data.timer.interval; }
-	set Interval(v) { this.setTimerInterval(v); }
-	get Enabled() { return this.data.timer.enabled; }
-	set Enabled(v) { this.setTimerEnabled(v); }
-}
+/**
+ * Amount of msecs to wait (at least) until same timer can be triggered again
+ * (e.g. they can fall behind, if set to > 1, as update cycle is 1000Hz)
+ */
+export const MAX_TIMER_MSEC_INTERVAL = 1;
+/**
+ * Amount of msecs that all timers combined can take per frame (e.g. they can
+ * fall behind, if set to < somelargevalue)
+ */
+export const MAX_TIMERS_MSEC_OVERALL = 5;
