+++ lib/vpt/timer/timer.ts
@@ -24,21 +24,12 @@ import { Storage } from '../../io/ole-doc';
 import { Item } from '../item';
 import { Table } from '../table/table';
 import { TimerData } from './timer-data';
-import { TimerHit } from './timer-hit';
+import { IScriptable } from '../../game/iscriptable';
+import { TimerApi } from './timer-api';
 
-/**
- * Amount of msecs to wait (at least) until same timer can be triggered again
- * (e.g. they can fall behind, if set to > 1, as update cycle is 1000Hz)
- */
-export const MAX_TIMER_MSEC_INTERVAL = 1;
+export class Timer extends Item<TimerData> implements IPlayable, IScriptable<TimerApi> {
 
-/**
- * Amount of msecs that all timers combined can take per frame (e.g. they can
- * fall behind, if set to < somelargevalue)
- */
-export const MAX_TIMERS_MSEC_OVERALL = 5;
-
-export class Timer extends Item<TimerData> implements IPlayable {
+	private api?: TimerApi;
 
 	public static async fromStorage(storage: Storage, itemName: string): Promise<Timer> {
 		const data = await TimerData.fromStorage(storage, itemName);
@@ -51,18 +42,15 @@ export class Timer extends Item<TimerData> implements IPlayable {
 
 	public setupPlayer(player: Player, table: Table): void {
 		this.events = new EventProxy(this);
+		this.api = new TimerApi(this.data, this.events, player, table);
 	}
 
 	public getEventNames(): string[] {
 		return [];
 	}
-}
 
-export class TimerOnOff {
-	public enabled: boolean;
-	public timer: TimerHit;
-	constructor(enabled: boolean, timer: TimerHit) {
-		this.enabled = enabled;
-		this.timer = timer;
+	public getApi(): TimerApi {
+		return this.api!;
 	}
 }
+
